refactor(cli): break the cli_dump_request import cycle instead of allowlisting it
Codex review on c335830 correctly flagged the prior fix: extending
IMPORT_CYCLE_ALLOWLIST for a freshly-discovered cycle needs an ADR or
explicit maintainer sign-off, not an inline comment (AGENTS.md's own
stated bar). Reverts that allowlist entry and fixes the actual
dependency direction instead.

cli_dump_request.py's only edge into the by-design CLI-registration SCC
was resolve_dump_request_for_cli's import of service_dump_pipeline
(TYPE_CHECKING for the return-type annotation, function-local for the
real call) -- everything else in the module (build_dump_request) only
ever imports api_types/compile_context. Moving that one function to
cli_buildsource.py (already a member of the SCC, already imported by
cli.py right next to it for _write_snapshot_output) makes
cli_dump_request.py a genuine leaf module with no path back into the
cluster, so no new SCC member forms at all -- not just an allowlisted
one.

Verified with the same stress test used to confirm the allowlist match
before reverting it: no cycle containing cli_dump_request across 50
randomized graph-traversal orderings (simulating the platform-dependent
DFS behavior that surfaced this on Windows but not Linux).
