+++ abicheck/cli_buildsource.py
@@ -73,7 +73,9 @@ from .cli_buildsource_helpers import (  # noqa: F401  (re-exported for API stabi
 )
 
 if TYPE_CHECKING:
+    from .api_types import DumpRequest
     from .model import AbiSnapshot
+    from .service_dump_pipeline import ResolvedDumpRequest
 
 
 # ── Attach / compare integration (ADR-028 D6, D7; ADR-029 D9) ─────────────────
@@ -704,6 +706,42 @@ def _write_snapshot_output(
         import json
 
         click.echo(json.dumps(payload, indent=2))
+
+
+def resolve_dump_request_for_cli(request: DumpRequest) -> ResolvedDumpRequest:
+    """:func:`~abicheck.service_dump_pipeline.resolve_dump_request`, with the
+    CLI's error contract.
+
+    The Tier-2 pipeline signals a bad request as
+    :class:`~abicheck.errors.ValidationError`; a Click front end owes the user
+    a ``UsageError`` (exit 64) instead. Translated here, at the boundary,
+    rather than inside the pipeline — the same Tier-1/Tier-2 separation
+    ``embed_side_build_source`` observes in the other direction.
+
+    Worth being explicit about what this can newly reject on the ``--dry-run``
+    path, since that path's own contract is "never raises on anything but a
+    usage error": :meth:`DumpRequest.validate` front-runs a check
+    ``dumper.dump()`` already performs at *runtime* — a ``--dump-manifest``
+    combined with ``-I``/``--include``, which declares two conflicting public
+    surfaces. That combination previously dry-ran as a clean report and then
+    failed during the real extraction. Reporting it as a usage error in both
+    places is a strictly better answer, and it stays inside the dry-run
+    contract because it is precisely a usage error.
+
+    Lives here, not in `cli_dump_request.py` (which builds the `DumpRequest`
+    this consumes), so that leaf module stays a leaf: this function's own
+    `service_dump_pipeline` import is the one edge that needs a module
+    already inside the by-design CLI-registration SCC
+    (`IMPORT_CYCLE_ALLOWLIST` in `scripts/check_ai_readiness.py`), which
+    `cli_buildsource` already is.
+    """
+    from .errors import ValidationError
+    from .service_dump_pipeline import resolve_dump_request
+
+    try:
+        return resolve_dump_request(request)
+    except ValidationError as exc:
+        raise click.UsageError(str(exc)) from exc
 
 
 # ── Back-compat re-export shim (lazy, to avoid an import cycle) ───────────────
